requirements.txt
prefect>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
yfinance>=0.2.0
alpaca-py>=0.8.0
websockets>=10.0
python-dotenv>=0.19.0 
streamlit~=1.45.1
pytz~=2025.2
loguru~=0.7.2
sqlalchemy~=2.0.23
redis~=6.2.0
newsapi-python=0.2.7